+++ api/CHANGELOG.md
@@ -5,9 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [unreleased]
+
+### Added
 
 - Add `ffi_derive` for `Measurement` and `Unit` (behind the `cffi` feature flag).
+- New feature `json-schema` implements `schemars::JsonSchema` for `Measurement`.
 
 ## [0.15.0] - 2020-07-31